Separating instructions onto their own files
