+++ src/System.Linq.Expressions/src/System/Linq/Expressions/Interpreter/ArrayOperations.cs
@@ -132,6 +132,24 @@ namespace System.Linq.Expressions.Interpreter
         }
     }
 
+    internal sealed class ArrayLengthInstruction : Instruction
+    {
+        public static readonly ArrayLengthInstruction Instance = new ArrayLengthInstruction();
+
+        public override int ConsumedStack => 1;
+        public override int ProducedStack => 1;
+        public override string InstructionName => "ArrayLength";
+
+        private ArrayLengthInstruction() { }
+
+        public override int Run(InterpretedFrame frame)
+        {
+            object obj = frame.Pop();
+            frame.Push(((Array)obj).Length);
+            return +1;
+        }
+    }
+
     internal sealed class ConvertHelper
     {
         public static int ToInt32NoNull(object val)
