+++ src/System.Linq.Expressions/src/System/Linq/Expressions/Interpreter/TypeOperations.cs
@@ -46,137 +46,6 @@ namespace System.Linq.Expressions.Interpreter
         }
     }
 
-    internal class NewInstruction : Instruction
-    {
-        protected readonly ConstructorInfo _constructor;
-        protected readonly int _argumentCount;
-
-        public NewInstruction(ConstructorInfo constructor, int argumentCount)
-        {
-            _constructor = constructor;
-            _argumentCount = argumentCount;
-        }
-
-        public override int ConsumedStack => _argumentCount;
-        public override int ProducedStack => 1;
-        public override string InstructionName => "New";
-
-        public override int Run(InterpretedFrame frame)
-        {
-            int first = frame.StackIndex - _argumentCount;
-
-            var args = GetArgs(frame, first);
-
-            object ret;
-            try
-            {
-                ret = _constructor.Invoke(args);
-            }
-            catch (TargetInvocationException e)
-            {
-                ExceptionHelpers.UpdateForRethrow(e.InnerException);
-                throw e.InnerException;
-            }
-
-            frame.Data[first] = ret;
-            frame.StackIndex = first + 1;
-
-            return +1;
-        }
-
-        protected object[] GetArgs(InterpretedFrame frame, int first)
-        {
-            if (_argumentCount > 0)
-            {
-                var args = new object[_argumentCount];
-
-                for (int i = 0; i < args.Length; i++)
-                {
-                    args[i] = frame.Data[first + i];
-                }
-
-                return args;
-            }
-            else
-            {
-                return Array.Empty<object>();
-            }
-        }
-
-        public override string ToString() => "New " + _constructor.DeclaringType.Name + "(" + _constructor + ")";
-    }
-
-    internal partial class ByRefNewInstruction : NewInstruction
-    {
-        private readonly ByRefUpdater[] _byrefArgs;
-
-        internal ByRefNewInstruction(ConstructorInfo target, int argumentCount, ByRefUpdater[] byrefArgs)
-            : base(target, argumentCount)
-        {
-            _byrefArgs = byrefArgs;
-        }
-
-        public override string InstructionName => "ByRefNew";
-        
-        public sealed override int Run(InterpretedFrame frame)
-        {
-            int first = frame.StackIndex - _argumentCount;
-
-            var args = GetArgs(frame, first);
-
-            try
-            {
-                object ret;
-                try
-                {
-                    ret = _constructor.Invoke(args);
-                }
-                catch (TargetInvocationException e)
-                {
-                    throw ExceptionHelpers.UpdateForRethrow(e.InnerException);
-                }
-
-                frame.Data[first] = ret;
-                frame.StackIndex = first + 1;
-            }
-            finally
-            {
-                if (args != null)
-                {
-                    foreach (var arg in _byrefArgs)
-                    {
-                        arg.Update(frame, args[arg.ArgumentIndex]);
-                    }
-                }
-            }
-
-            return 1;
-        }
-    }
-
-    internal sealed class DefaultValueInstruction : Instruction
-    {
-        private readonly Type _type;
-
-        internal DefaultValueInstruction(Type type)
-        {
-            _type = type;
-        }
-
-        public override int ConsumedStack => 0;
-        public override int ProducedStack => 1;
-        public override string InstructionName => "DefaultValue";
-        
-        public override int Run(InterpretedFrame frame)
-        {
-            object value = _type.GetTypeInfo().IsValueType ? Activator.CreateInstance(_type) : null;
-            frame.Push(value);
-            return +1;
-        }
-
-        public override string ToString() => "New " + _type;
-    }
-
     internal sealed class TypeIsInstruction : Instruction
     {
         private readonly Type _type;
@@ -264,1657 +133,6 @@ namespace System.Linq.Expressions.Interpreter
             object obj = frame.Pop();
             frame.Push(ScriptingRuntimeHelpers.BooleanToObject(obj != null && (object)obj.GetType() == type));
             return +1;
-        }
-    }
-
-    internal sealed class ArrayLengthInstruction : Instruction
-    {
-        public static readonly ArrayLengthInstruction Instance = new ArrayLengthInstruction();
-
-        public override int ConsumedStack => 1;
-        public override int ProducedStack => 1;
-        public override string InstructionName => "ArrayLength";
-        
-        private ArrayLengthInstruction() { }
-
-        public override int Run(InterpretedFrame frame)
-        {
-            object obj = frame.Pop();
-            frame.Push(((Array)obj).Length);
-            return +1;
-        }
-    }
-
-    internal abstract class NegateInstruction : Instruction
-    {
-        private static Instruction s_int16, s_int32, s_int64, s_single, s_double;
-
-        public override int ConsumedStack => 1;
-        public override int ProducedStack => 1;
-        public override string InstructionName => "Negate";
-        
-        private NegateInstruction() { }
-
-        internal sealed class NegateInt32 : NegateInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(ScriptingRuntimeHelpers.Int32ToObject(unchecked(-(Int32)obj)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class NegateInt16 : NegateInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(unchecked((Int16)(-(Int16)obj)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class NegateInt64 : NegateInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(unchecked((Int64)(-(Int64)obj)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class NegateSingle : NegateInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(unchecked((Single)(-(Single)obj)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class NegateDouble : NegateInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(unchecked((Double)(-(Double)obj)));
-                }
-                return +1;
-            }
-        }
-
-        public static Instruction Create(Type type)
-        {
-            Debug.Assert(!type.GetTypeInfo().IsEnum);
-            switch (System.Dynamic.Utils.TypeExtensions.GetTypeCode(TypeUtils.GetNonNullableType(type)))
-            {
-                case TypeCode.Int16: return s_int16 ?? (s_int16 = new NegateInt16());
-                case TypeCode.Int32: return s_int32 ?? (s_int32 = new NegateInt32());
-                case TypeCode.Int64: return s_int64 ?? (s_int64 = new NegateInt64());
-                case TypeCode.Single: return s_single ?? (s_single = new NegateSingle());
-                case TypeCode.Double: return s_double ?? (s_double = new NegateDouble());
-
-                default:
-                    throw Error.ExpressionNotSupportedForType("Negate", type);
-            }
-        }
-    }
-
-    internal abstract class NegateCheckedInstruction : Instruction
-    {
-        private static Instruction s_int16, s_int32, s_int64, s_single, s_double;
-
-        public override int ConsumedStack => 1;
-        public override int ProducedStack => 1;
-        public override string InstructionName => "NegateChecked";
-        
-        private NegateCheckedInstruction() { }
-
-        internal sealed class NegateCheckedInt32 : NegateCheckedInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(ScriptingRuntimeHelpers.Int32ToObject(checked(-(Int32)obj)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class NegateCheckedInt16 : NegateCheckedInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(checked((Int16)(-(Int16)obj)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class NegateCheckedInt64 : NegateCheckedInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(checked((Int64)(-(Int64)obj)));
-                }
-                return +1;
-            }
-        }
-        internal sealed class NegateCheckedSingle : NegateCheckedInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(checked((Single)(-(Single)obj)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class NegateCheckedDouble : NegateCheckedInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(checked((Double)(-(Double)obj)));
-                }
-                return +1;
-            }
-        }
-
-        public static Instruction Create(Type type)
-        {
-            Debug.Assert(!type.GetTypeInfo().IsEnum);
-            switch (System.Dynamic.Utils.TypeExtensions.GetTypeCode(TypeUtils.GetNonNullableType(type)))
-            {
-                case TypeCode.Int16: return s_int16 ?? (s_int16 = new NegateCheckedInt16());
-                case TypeCode.Int32: return s_int32 ?? (s_int32 = new NegateCheckedInt32());
-                case TypeCode.Int64: return s_int64 ?? (s_int64 = new NegateCheckedInt64());
-                case TypeCode.Single: return s_single ?? (s_single = new NegateCheckedSingle());
-                case TypeCode.Double: return s_double ?? (s_double = new NegateCheckedDouble());
-                default:
-                    throw Error.ExpressionNotSupportedForType("NegateChecked", type);
-            }
-        }
-    }
-
-    internal abstract class OnesComplementInstruction : Instruction
-    {
-        private static Instruction s_byte, s_sbyte, s_int16, s_int32, s_int64, s_UInt16, s_UInt32, s_UInt64;
-
-        public override int ConsumedStack => 1;
-        public override int ProducedStack => 1;
-        public override string InstructionName => "OnesComplement";
-
-        private OnesComplementInstruction() { }
-
-        internal sealed class OnesComplementInt32 : OnesComplementInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(ScriptingRuntimeHelpers.Int32ToObject(~(Int32)obj));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class OnesComplementInt16 : OnesComplementInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push((Int16)(~(Int16)obj));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class OnesComplementInt64 : OnesComplementInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push((Int64)(~(Int64)obj));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class OnesComplementUInt16 : OnesComplementInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push((UInt16)(~(UInt16)obj));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class OnesComplementUInt32 : OnesComplementInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push((UInt32)(~(UInt32)obj));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class OnesComplementUInt64 : OnesComplementInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push((UInt64)(~(UInt64)obj));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class OnesComplementByte : OnesComplementInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push((Byte)(~(Byte)obj));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class OnesComplementSByte : OnesComplementInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push((SByte)(~(SByte)obj));
-                }
-                return +1;
-            }
-        }
-
-        public static Instruction Create(Type type)
-        {
-            Debug.Assert(!type.GetTypeInfo().IsEnum);
-            switch (System.Dynamic.Utils.TypeExtensions.GetTypeCode(TypeUtils.GetNonNullableType(type)))
-            {
-                case TypeCode.Byte: return s_byte ?? (s_byte = new OnesComplementByte());
-                case TypeCode.SByte: return s_sbyte ?? (s_sbyte = new OnesComplementSByte());
-                case TypeCode.Int16: return s_int16 ?? (s_int16 = new OnesComplementInt16());
-                case TypeCode.Int32: return s_int32 ?? (s_int32 = new OnesComplementInt32());
-                case TypeCode.Int64: return s_int64 ?? (s_int64 = new OnesComplementInt64());
-                case TypeCode.UInt16: return s_UInt16 ?? (s_UInt16 = new OnesComplementUInt16());
-                case TypeCode.UInt32: return s_UInt32 ?? (s_UInt32 = new OnesComplementUInt32());
-                case TypeCode.UInt64: return s_UInt64 ?? (s_UInt64 = new OnesComplementUInt64());
-
-                default:
-                    throw Error.ExpressionNotSupportedForType("OnesComplement", type);
-            }
-        }
-    }
-
-    internal abstract class IncrementInstruction : Instruction
-    {
-        private static Instruction s_int16, s_int32, s_int64, s_UInt16, s_UInt32, s_UInt64, s_single, s_double;
-
-        public override int ConsumedStack => 1;
-        public override int ProducedStack => 1;
-        public override string InstructionName => "Increment";
-        
-        private IncrementInstruction() { }
-
-        internal sealed class IncrementInt32 : IncrementInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(ScriptingRuntimeHelpers.Int32ToObject(unchecked(1 + (Int32)obj)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class IncrementInt16 : IncrementInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(unchecked((Int16)(1 + (Int16)obj)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class IncrementInt64 : IncrementInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(unchecked((Int64)(1 + (Int64)obj)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class IncrementUInt16 : IncrementInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(unchecked((UInt16)(1 + (UInt16)obj)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class IncrementUInt32 : IncrementInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(unchecked((UInt32)(1 + (UInt32)obj)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class IncrementUInt64 : IncrementInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(unchecked((UInt64)(1 + (UInt64)obj)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class IncrementSingle : IncrementInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(unchecked((Single)(1 + (Single)obj)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class IncrementDouble : IncrementInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(unchecked((Double)(1 + (Double)obj)));
-                }
-                return +1;
-            }
-        }
-
-        public static Instruction Create(Type type)
-        {
-            Debug.Assert(!type.GetTypeInfo().IsEnum);
-            switch (System.Dynamic.Utils.TypeExtensions.GetTypeCode(TypeUtils.GetNonNullableType(type)))
-            {
-                case TypeCode.Int16: return s_int16 ?? (s_int16 = new IncrementInt16());
-                case TypeCode.Int32: return s_int32 ?? (s_int32 = new IncrementInt32());
-                case TypeCode.Int64: return s_int64 ?? (s_int64 = new IncrementInt64());
-                case TypeCode.UInt16: return s_UInt16 ?? (s_UInt16 = new IncrementUInt16());
-                case TypeCode.UInt32: return s_UInt32 ?? (s_UInt32 = new IncrementUInt32());
-                case TypeCode.UInt64: return s_UInt64 ?? (s_UInt64 = new IncrementUInt64());
-                case TypeCode.Single: return s_single ?? (s_single = new IncrementSingle());
-                case TypeCode.Double: return s_double ?? (s_double = new IncrementDouble());
-
-                default:
-                    throw Error.ExpressionNotSupportedForType("Increment", type);
-            }
-        }
-    }
-
-    internal abstract class DecrementInstruction : Instruction
-    {
-        private static Instruction s_int16, s_int32, s_int64, s_UInt16, s_UInt32, s_UInt64, s_single, s_double;
-
-        public override int ConsumedStack => 1;
-        public override int ProducedStack => 1;
-        public override string InstructionName => "Decrement";
-
-        private DecrementInstruction() { }
-
-        internal sealed class DecrementInt32 : DecrementInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(ScriptingRuntimeHelpers.Int32ToObject(unchecked((Int32)obj - 1)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class DecrementInt16 : DecrementInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(unchecked((Int16)((Int16)obj - 1)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class DecrementInt64 : DecrementInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(unchecked((Int64)((Int64)obj - 1)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class DecrementUInt16 : DecrementInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(unchecked((UInt16)((UInt16)obj - 1)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class DecrementUInt32 : DecrementInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(unchecked((UInt32)((UInt32)obj - 1)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class DecrementUInt64 : DecrementInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(unchecked((UInt64)((UInt64)obj - 1)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class DecrementSingle : DecrementInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(unchecked((Single)((Single)obj - 1)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class DecrementDouble : DecrementInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                object obj = frame.Pop();
-                if (obj == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(unchecked((Double)((Double)obj - 1)));
-                }
-                return +1;
-            }
-        }
-
-        public static Instruction Create(Type type)
-        {
-            Debug.Assert(!type.GetTypeInfo().IsEnum);
-            switch (System.Dynamic.Utils.TypeExtensions.GetTypeCode(TypeUtils.GetNonNullableType(type)))
-            {
-                case TypeCode.Int16: return s_int16 ?? (s_int16 = new DecrementInt16());
-                case TypeCode.Int32: return s_int32 ?? (s_int32 = new DecrementInt32());
-                case TypeCode.Int64: return s_int64 ?? (s_int64 = new DecrementInt64());
-                case TypeCode.UInt16: return s_UInt16 ?? (s_UInt16 = new DecrementUInt16());
-                case TypeCode.UInt32: return s_UInt32 ?? (s_UInt32 = new DecrementUInt32());
-                case TypeCode.UInt64: return s_UInt64 ?? (s_UInt64 = new DecrementUInt64());
-                case TypeCode.Single: return s_single ?? (s_single = new DecrementSingle());
-                case TypeCode.Double: return s_double ?? (s_double = new DecrementDouble());
-
-                default:
-                    throw Error.ExpressionNotSupportedForType("Decrement", type);
-            }
-        }
-    }
-
-
-    internal abstract class LeftShiftInstruction : Instruction
-    {
-        private static Instruction s_SByte, s_int16, s_int32, s_int64, s_byte, s_UInt16, s_UInt32, s_UInt64;
-
-        public override int ConsumedStack => 2;
-        public override int ProducedStack => 1;
-        public override string InstructionName => "LeftShift";
-        
-        private LeftShiftInstruction() { }
-
-        internal sealed class LeftShiftSByte : LeftShiftInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var shift = frame.Pop();
-                var value = frame.Pop();
-                if (value == null || shift == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push((SByte)(((SByte)value) << ((int)shift)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class LeftShiftInt16 : LeftShiftInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var shift = frame.Pop();
-                var value = frame.Pop();
-                if (value == null || shift == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push((Int16)(((Int16)value) << ((int)shift)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class LeftShiftInt32 : LeftShiftInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var shift = frame.Pop();
-                var value = frame.Pop();
-                if (value == null || shift == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(((Int32)value) << ((int)shift));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class LeftShiftInt64 : LeftShiftInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var shift = frame.Pop();
-                var value = frame.Pop();
-                if (value == null || shift == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(((Int64)value) << ((int)shift));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class LeftShiftByte : LeftShiftInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var shift = frame.Pop();
-                var value = frame.Pop();
-                if (value == null || shift == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push((Byte)(((Byte)value) << ((int)shift)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class LeftShiftUInt16 : LeftShiftInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var shift = frame.Pop();
-                var value = frame.Pop();
-                if (value == null || shift == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push((UInt16)(((UInt16)value) << ((int)shift)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class LeftShiftUInt32 : LeftShiftInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var shift = frame.Pop();
-                var value = frame.Pop();
-                if (value == null || shift == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(((UInt32)value) << ((int)shift));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class LeftShiftUInt64 : LeftShiftInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var shift = frame.Pop();
-                var value = frame.Pop();
-                if (value == null || shift == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(((UInt64)value) << ((int)shift));
-                }
-                return +1;
-            }
-        }
-
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Maintainability", "CA1502:AvoidExcessiveComplexity")]
-        public static Instruction Create(Type type)
-        {
-            // Boxed enums can be unboxed as their underlying types:
-            switch (System.Dynamic.Utils.TypeExtensions.GetTypeCode(type.GetTypeInfo().IsEnum ? Enum.GetUnderlyingType(type) : TypeUtils.GetNonNullableType(type)))
-            {
-                case TypeCode.SByte: return s_SByte ?? (s_SByte = new LeftShiftSByte());
-                case TypeCode.Byte: return s_byte ?? (s_byte = new LeftShiftByte());
-                case TypeCode.Int16: return s_int16 ?? (s_int16 = new LeftShiftInt16());
-                case TypeCode.Int32: return s_int32 ?? (s_int32 = new LeftShiftInt32());
-                case TypeCode.Int64: return s_int64 ?? (s_int64 = new LeftShiftInt64());
-
-                case TypeCode.UInt16: return s_UInt16 ?? (s_UInt16 = new LeftShiftUInt16());
-                case TypeCode.UInt32: return s_UInt32 ?? (s_UInt32 = new LeftShiftUInt32());
-                case TypeCode.UInt64: return s_UInt64 ?? (s_UInt64 = new LeftShiftUInt64());
-
-                default:
-                    throw Error.ExpressionNotSupportedForType("LeftShift", type);
-            }
-        }
-    }
-
-    internal abstract class RightShiftInstruction : Instruction
-    {
-        private static Instruction s_SByte, s_int16, s_int32, s_int64, s_byte, s_UInt16, s_UInt32, s_UInt64;
-
-        public override int ConsumedStack => 2;
-        public override int ProducedStack => 1;
-        public override string InstructionName => "RightShift";
-        
-        private RightShiftInstruction() { }
-
-        internal sealed class RightShiftSByte : RightShiftInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var shift = frame.Pop();
-                var value = frame.Pop();
-                if (value == null || shift == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push((SByte)(((SByte)value) >> ((int)shift)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class RightShiftInt16 : RightShiftInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var shift = frame.Pop();
-                var value = frame.Pop();
-                if (value == null || shift == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push((Int16)(((Int16)value) >> ((int)shift)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class RightShiftInt32 : RightShiftInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var shift = frame.Pop();
-                var value = frame.Pop();
-                if (value == null || shift == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(((Int32)value) >> ((int)shift));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class RightShiftInt64 : RightShiftInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var shift = frame.Pop();
-                var value = frame.Pop();
-                if (value == null || shift == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(((Int64)value) >> ((int)shift));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class RightShiftByte : RightShiftInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var shift = frame.Pop();
-                var value = frame.Pop();
-                if (value == null || shift == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push((Byte)(((Byte)value) >> ((int)shift)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class RightShiftUInt16 : RightShiftInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var shift = frame.Pop();
-                var value = frame.Pop();
-                if (value == null || shift == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push((UInt16)(((UInt16)value) >> ((int)shift)));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class RightShiftUInt32 : RightShiftInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var shift = frame.Pop();
-                var value = frame.Pop();
-                if (value == null || shift == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(((UInt32)value) >> ((int)shift));
-                }
-                return +1;
-            }
-        }
-
-        internal sealed class RightShiftUInt64 : RightShiftInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var shift = frame.Pop();
-                var value = frame.Pop();
-                if (value == null || shift == null)
-                {
-                    frame.Push(null);
-                }
-                else
-                {
-                    frame.Push(((UInt64)value) >> ((int)shift));
-                }
-                return +1;
-            }
-        }
-
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Maintainability", "CA1502:AvoidExcessiveComplexity")]
-        public static Instruction Create(Type type)
-        {
-            // Boxed enums can be unboxed as their underlying types:
-            switch (System.Dynamic.Utils.TypeExtensions.GetTypeCode(type.GetTypeInfo().IsEnum ? Enum.GetUnderlyingType(type) : TypeUtils.GetNonNullableType(type)))
-            {
-                case TypeCode.SByte: return s_SByte ?? (s_SByte = new RightShiftSByte());
-                case TypeCode.Byte: return s_byte ?? (s_byte = new RightShiftByte());
-                case TypeCode.Int16: return s_int16 ?? (s_int16 = new RightShiftInt16());
-                case TypeCode.Int32: return s_int32 ?? (s_int32 = new RightShiftInt32());
-                case TypeCode.Int64: return s_int64 ?? (s_int64 = new RightShiftInt64());
-
-                case TypeCode.UInt16: return s_UInt16 ?? (s_UInt16 = new RightShiftUInt16());
-                case TypeCode.UInt32: return s_UInt32 ?? (s_UInt32 = new RightShiftUInt32());
-                case TypeCode.UInt64: return s_UInt64 ?? (s_UInt64 = new RightShiftUInt64());
-
-                default:
-                    throw Error.ExpressionNotSupportedForType("RightShift", type);
-            }
-        }
-    }
-
-    internal abstract class ExclusiveOrInstruction : Instruction
-    {
-        private static Instruction s_SByte, s_int16, s_int32, s_int64, s_byte, s_UInt16, s_UInt32, s_UInt64, s_bool;
-
-        public override int ConsumedStack => 2;
-        public override int ProducedStack => 1;
-        public override string InstructionName => "ExclusiveOr";
-        
-        private ExclusiveOrInstruction() { }
-
-        internal sealed class ExclusiveOrSByte : ExclusiveOrInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var left = frame.Pop();
-                var right = frame.Pop();
-                if (left == null || right == null)
-                {
-                    frame.Push(null);
-                    return +1;
-                }
-                frame.Push((SByte)(((SByte)left) ^ ((SByte)right)));
-                return +1;
-            }
-        }
-
-        internal sealed class ExclusiveOrInt16 : ExclusiveOrInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var left = frame.Pop();
-                var right = frame.Pop();
-                if (left == null || right == null)
-                {
-                    frame.Push(null);
-                    return +1;
-                }
-                frame.Push((Int16)(((Int16)left) ^ ((Int16)right)));
-                return +1;
-            }
-        }
-
-        internal sealed class ExclusiveOrInt32 : ExclusiveOrInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var left = frame.Pop();
-                var right = frame.Pop();
-                if (left == null || right == null)
-                {
-                    frame.Push(null);
-                    return +1;
-                }
-                frame.Push(((Int32)left) ^ ((Int32)right));
-                return +1;
-            }
-        }
-
-        internal sealed class ExclusiveOrInt64 : ExclusiveOrInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var left = frame.Pop();
-                var right = frame.Pop();
-                if (left == null || right == null)
-                {
-                    frame.Push(null);
-                    return +1;
-                }
-                frame.Push(((Int64)left) ^ ((Int64)right));
-                return +1;
-            }
-        }
-
-        internal sealed class ExclusiveOrByte : ExclusiveOrInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var left = frame.Pop();
-                var right = frame.Pop();
-                if (left == null || right == null)
-                {
-                    frame.Push(null);
-                    return +1;
-                }
-                frame.Push((Byte)(((Byte)left) ^ ((Byte)right)));
-                return +1;
-            }
-        }
-
-        internal sealed class ExclusiveOrUInt16 : ExclusiveOrInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var left = frame.Pop();
-                var right = frame.Pop();
-                if (left == null || right == null)
-                {
-                    frame.Push(null);
-                    return +1;
-                }
-                frame.Push((UInt16)(((UInt16)left) ^ ((UInt16)right)));
-                return +1;
-            }
-        }
-
-        internal sealed class ExclusiveOrUInt32 : ExclusiveOrInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var left = frame.Pop();
-                var right = frame.Pop();
-                if (left == null || right == null)
-                {
-                    frame.Push(null);
-                    return +1;
-                }
-                frame.Push(((UInt32)left) ^ ((UInt32)right));
-                return +1;
-            }
-        }
-
-        internal sealed class ExclusiveOrUInt64 : ExclusiveOrInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var left = frame.Pop();
-                var right = frame.Pop();
-                if (left == null || right == null)
-                {
-                    frame.Push(null);
-                    return +1;
-                }
-                frame.Push(((UInt64)left) ^ ((UInt64)right));
-                return +1;
-            }
-        }
-
-        internal sealed class ExclusiveOrBool : ExclusiveOrInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var left = frame.Pop();
-                var right = frame.Pop();
-                if (left == null || right == null)
-                {
-                    frame.Push(null);
-                    return +1;
-                }
-                frame.Push(((Boolean)left) ^ ((Boolean)right));
-                return +1;
-            }
-        }
-
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Maintainability", "CA1502:AvoidExcessiveComplexity")]
-        public static Instruction Create(Type type)
-        {
-            // Boxed enums can be unboxed as their underlying types:
-            switch (GetTypeCode(type))
-            {
-                case TypeCode.SByte: return s_SByte ?? (s_SByte = new ExclusiveOrSByte());
-                case TypeCode.Byte: return s_byte ?? (s_byte = new ExclusiveOrByte());
-                case TypeCode.Int16: return s_int16 ?? (s_int16 = new ExclusiveOrInt16());
-                case TypeCode.Int32: return s_int32 ?? (s_int32 = new ExclusiveOrInt32());
-                case TypeCode.Int64: return s_int64 ?? (s_int64 = new ExclusiveOrInt64());
-
-                case TypeCode.UInt16: return s_UInt16 ?? (s_UInt16 = new ExclusiveOrUInt16());
-                case TypeCode.UInt32: return s_UInt32 ?? (s_UInt32 = new ExclusiveOrUInt32());
-                case TypeCode.UInt64: return s_UInt64 ?? (s_UInt64 = new ExclusiveOrUInt64());
-                case TypeCode.Boolean: return s_bool ?? (s_bool = new ExclusiveOrBool());
-
-                default:
-                    throw Error.ExpressionNotSupportedForType("ExclusiveOr", type);
-            }
-        }
-
-        private static TypeCode GetTypeCode(Type type)
-        {
-            return System.Dynamic.Utils.TypeExtensions.GetTypeCode(type.GetTypeInfo().IsEnum ? Enum.GetUnderlyingType(type) : TypeUtils.GetNonNullableType(type));
-        }
-    }
-
-    internal abstract class OrInstruction : Instruction
-    {
-        private static Instruction s_SByte, s_int16, s_int32, s_int64, s_byte, s_UInt16, s_UInt32, s_UInt64, s_bool;
-
-        public override int ConsumedStack => 2;
-        public override int ProducedStack => 1;
-        public override string InstructionName => "Or";
-        
-        private OrInstruction() { }
-
-        internal sealed class OrSByte : OrInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var left = frame.Pop();
-                var right = frame.Pop();
-                if (left == null || right == null)
-                {
-                    frame.Push(null);
-                    return +1;
-                }
-                frame.Push((SByte)(((SByte)left) | ((SByte)right)));
-                return +1;
-            }
-        }
-
-        internal sealed class OrInt16 : OrInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var left = frame.Pop();
-                var right = frame.Pop();
-                if (left == null || right == null)
-                {
-                    frame.Push(null);
-                    return +1;
-                }
-                frame.Push((Int16)(((Int16)left) | ((Int16)right)));
-                return +1;
-            }
-        }
-
-        internal sealed class OrInt32 : OrInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var left = frame.Pop();
-                var right = frame.Pop();
-                if (left == null || right == null)
-                {
-                    frame.Push(null);
-                    return +1;
-                }
-                frame.Push(((Int32)left) | ((Int32)right));
-                return +1;
-            }
-        }
-
-        internal sealed class OrInt64 : OrInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var left = frame.Pop();
-                var right = frame.Pop();
-                if (left == null || right == null)
-                {
-                    frame.Push(null);
-                    return +1;
-                }
-                frame.Push(((Int64)left) | ((Int64)right));
-                return +1;
-            }
-        }
-
-        internal sealed class OrByte : OrInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var left = frame.Pop();
-                var right = frame.Pop();
-                if (left == null || right == null)
-                {
-                    frame.Push(null);
-                    return +1;
-                }
-                frame.Push((Byte)(((Byte)left) | ((Byte)right)));
-                return +1;
-            }
-        }
-
-        internal sealed class OrUInt16 : OrInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var left = frame.Pop();
-                var right = frame.Pop();
-                if (left == null || right == null)
-                {
-                    frame.Push(null);
-                    return +1;
-                }
-                frame.Push((UInt16)(((UInt16)left) | ((UInt16)right)));
-                return +1;
-            }
-        }
-
-        internal sealed class OrUInt32 : OrInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var left = frame.Pop();
-                var right = frame.Pop();
-                if (left == null || right == null)
-                {
-                    frame.Push(null);
-                    return +1;
-                }
-                frame.Push(((UInt32)left) | ((UInt32)right));
-                return +1;
-            }
-        }
-
-        internal sealed class OrUInt64 : OrInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var left = frame.Pop();
-                var right = frame.Pop();
-                if (left == null || right == null)
-                {
-                    frame.Push(null);
-                    return +1;
-                }
-                frame.Push(((UInt64)left) | ((UInt64)right));
-                return +1;
-            }
-        }
-
-        internal sealed class OrBool : OrInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var right = frame.Pop();
-                var left = frame.Pop();
-                if (left == null)
-                {
-                    if (right == null)
-                    {
-                        frame.Push(null);
-                    }
-                    else
-                    {
-                        frame.Push((Boolean)right ? ScriptingRuntimeHelpers.True : null);
-                    }
-                    return +1;
-                }
-                else if (right == null)
-                {
-                    frame.Push((Boolean)left ? ScriptingRuntimeHelpers.True : null);
-                    return +1;
-                }
-                frame.Push(((Boolean)left) | ((Boolean)right));
-                return +1;
-            }
-        }
-
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Maintainability", "CA1502:AvoidExcessiveComplexity")]
-        public static Instruction Create(Type type)
-        {
-            // Boxed enums can be unboxed as their underlying types:
-            switch (System.Dynamic.Utils.TypeExtensions.GetTypeCode(type.GetTypeInfo().IsEnum ? Enum.GetUnderlyingType(type) : TypeUtils.GetNonNullableType(type)))
-            {
-                case TypeCode.SByte: return s_SByte ?? (s_SByte = new OrSByte());
-                case TypeCode.Byte: return s_byte ?? (s_byte = new OrByte());
-                case TypeCode.Int16: return s_int16 ?? (s_int16 = new OrInt16());
-                case TypeCode.Int32: return s_int32 ?? (s_int32 = new OrInt32());
-                case TypeCode.Int64: return s_int64 ?? (s_int64 = new OrInt64());
-
-                case TypeCode.UInt16: return s_UInt16 ?? (s_UInt16 = new OrUInt16());
-                case TypeCode.UInt32: return s_UInt32 ?? (s_UInt32 = new OrUInt32());
-                case TypeCode.UInt64: return s_UInt64 ?? (s_UInt64 = new OrUInt64());
-                case TypeCode.Boolean: return s_bool ?? (s_bool = new OrBool());
-
-                default:
-                    throw Error.ExpressionNotSupportedForType("Or", type);
-            }
-        }
-    }
-
-    internal abstract class AndInstruction : Instruction
-    {
-        private static Instruction s_SByte, s_int16, s_int32, s_int64, s_byte, s_UInt16, s_UInt32, s_UInt64, s_bool;
-
-        public override int ConsumedStack => 2;
-        public override int ProducedStack => 1;
-        public override string InstructionName => "And";
-        
-        private AndInstruction() { }
-
-        internal sealed class AndSByte : AndInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var left = frame.Pop();
-                var right = frame.Pop();
-                if (left == null || right == null)
-                {
-                    frame.Push(null);
-                    return +1;
-                }
-                frame.Push((SByte)(((SByte)left) & ((SByte)right)));
-                return +1;
-            }
-        }
-
-        internal sealed class AndInt16 : AndInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var left = frame.Pop();
-                var right = frame.Pop();
-                if (left == null || right == null)
-                {
-                    frame.Push(null);
-                    return +1;
-                }
-                frame.Push((Int16)(((Int16)left) & ((Int16)right)));
-                return +1;
-            }
-        }
-
-        internal sealed class AndInt32 : AndInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var left = frame.Pop();
-                var right = frame.Pop();
-                if (left == null || right == null)
-                {
-                    frame.Push(null);
-                    return +1;
-                }
-                frame.Push(((Int32)left) & ((Int32)right));
-                return +1;
-            }
-        }
-
-        internal sealed class AndInt64 : AndInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var left = frame.Pop();
-                var right = frame.Pop();
-                if (left == null || right == null)
-                {
-                    frame.Push(null);
-                    return +1;
-                }
-                frame.Push(((Int64)left) & ((Int64)right));
-                return +1;
-            }
-        }
-
-        internal sealed class AndByte : AndInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var left = frame.Pop();
-                var right = frame.Pop();
-                if (left == null || right == null)
-                {
-                    frame.Push(null);
-                    return +1;
-                }
-                frame.Push((Byte)(((Byte)left) & ((Byte)right)));
-                return +1;
-            }
-        }
-
-        internal sealed class AndUInt16 : AndInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var left = frame.Pop();
-                var right = frame.Pop();
-                if (left == null || right == null)
-                {
-                    frame.Push(null);
-                    return +1;
-                }
-                frame.Push((UInt16)(((UInt16)left) & ((UInt16)right)));
-                return +1;
-            }
-        }
-
-        internal sealed class AndUInt32 : AndInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var left = frame.Pop();
-                var right = frame.Pop();
-                if (left == null || right == null)
-                {
-                    frame.Push(null);
-                    return +1;
-                }
-                frame.Push(((UInt32)left) & ((UInt32)right));
-                return +1;
-            }
-        }
-
-        internal sealed class AndUInt64 : AndInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var left = frame.Pop();
-                var right = frame.Pop();
-                if (left == null || right == null)
-                {
-                    frame.Push(null);
-                    return +1;
-                }
-                frame.Push(((UInt64)left) & ((UInt64)right));
-                return +1;
-            }
-        }
-
-        internal sealed class AndBool : AndInstruction
-        {
-            public override int Run(InterpretedFrame frame)
-            {
-                var right = frame.Pop();
-                var left = frame.Pop();
-                if (left == null)
-                {
-                    if (right == null)
-                    {
-                        frame.Push(null);
-                    }
-                    else
-                    {
-                        frame.Push((Boolean)right ? null : ScriptingRuntimeHelpers.False);
-                    }
-                    return +1;
-                }
-                else if (right == null)
-                {
-                    frame.Push((Boolean)left ? null : ScriptingRuntimeHelpers.False);
-                    return +1;
-                }
-                frame.Push(((Boolean)left) & ((Boolean)right));
-                return +1;
-            }
-        }
-
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Maintainability", "CA1502:AvoidExcessiveComplexity")]
-        public static Instruction Create(Type type)
-        {
-            // Boxed enums can be unboxed as their underlying types:
-            switch (System.Dynamic.Utils.TypeExtensions.GetTypeCode(type.GetTypeInfo().IsEnum ? Enum.GetUnderlyingType(type) : TypeUtils.GetNonNullableType(type)))
-            {
-                case TypeCode.SByte: return s_SByte ?? (s_SByte = new AndSByte());
-                case TypeCode.Byte: return s_byte ?? (s_byte = new AndByte());
-                case TypeCode.Int16: return s_int16 ?? (s_int16 = new AndInt16());
-                case TypeCode.Int32: return s_int32 ?? (s_int32 = new AndInt32());
-                case TypeCode.Int64: return s_int64 ?? (s_int64 = new AndInt64());
-
-                case TypeCode.UInt16: return s_UInt16 ?? (s_UInt16 = new AndUInt16());
-                case TypeCode.UInt32: return s_UInt32 ?? (s_UInt32 = new AndUInt32());
-                case TypeCode.UInt64: return s_UInt64 ?? (s_UInt64 = new AndUInt64());
-                case TypeCode.Boolean: return s_bool ?? (s_bool = new AndBool());
-
-                default:
-                    throw Error.ExpressionNotSupportedForType("And", type);
-            }
         }
     }
